Change measure width to be decimal instead of int (fixes #5)

+++ MusicXml/Domain/Measure.cs
@@ -10,7 +10,7 @@ namespace MusicXml.Domain
 			MeasureElements = new List<MeasureElement>();
 		}
 
-		public int Width { get; internal set; }
+		public decimal Width { get; internal set; }
 		
 		// This can be any musicXML element in the measure tag, i.e. note, backup, etc
 		public List<MeasureElement> MeasureElements { get; internal set; }
